Small bugfixes
Redraw the screen immediately on right click press/release
High playback speeds skip frames instead of maxing out at redraw speed

+++ src/sdlapp.h
@@ -37,7 +37,7 @@ struct App
     int width, height;
     struct Viewer view;
     GraphicList images, current_frame;
-    size_t timer;
+    double timer;
     bool paused, looping;
     double playback_speed;
     bool state_text_visible;
